add redis and implement some crud

+++ CourseService/src/main/java/com/example/courseservice/model/UserCourseHistory.java
@@ -20,11 +20,11 @@ public class UserCourseHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     // Reference to user-service
     @Column(name = "user_id", nullable = false)
-    private long userId;
+    private Integer userId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "course_id", nullable = false)
